fix(issue-433): Address PR review - harden hook ownership
